% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging.R
\name{bagging}
\alias{bagging}
\title{Bagging Algorithm (Regression or Classification)}
\usage{
bagging(data, b = 5, type = NULL, pred_val = NULL)
}
\arguments{
\item{data}{a named list that contains Regression or Classification data\cr the x values have the name x and are in
the form of a matrix \cr where the row number gives the dimension of the data\cr the y
values have the name y and are in the form of a vector}

\item{b}{Number of Bags to create. The Default Value is 5.}

\item{type}{Whether Classification or Regression is being called}

\item{pred_val}{OPTIONAL: If wanting to predict some values in a matrix, each of length \cr equal one less than
the number of dimensions in our data set. \cr Else, input null.}
}
\value{
A named list with the elements \code{Bagged_Trees} and \code{Prediction}.\cr
\code{Bagged_Trees} is an unnamed list with a number of tree \cr elements (see ?greedy_cart_regression) equal to the input value \code{b}
\code{Prediction} is the predicted value resulting from the \code{pred_val} data point. \cr If no \code{pred_val} was provided, this value is NULL
}
\description{
Bagging Algorithm for Regression or Classification Data
}
\examples{
bagging(create_random_sample_data_reg_dim(5, 100, 4), 3, type = "reg", matrix(c(0.1, 0.4, 0.6, 0.2, 0.3, 0.6), ncol = 2))
bagging(create_Sample_data_class(1), 4, type = "class", matrix(c(0.1,0.4,0.6, 0.2, 0.4, 0, 1, 1), ncol = 4))
}
