% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruning.R
\name{count_leaves}
\alias{count_leaves}
\title{Count leaves for a given CART}
\usage{
count_leaves(data)
}
\arguments{
\item{data}{A tibble containing the information of each node type of the CART in the colum "name".}
}
\value{
The number of nodes titled leaf of the CART \code{data}
}
\description{
Count leaves for a given CART
}
\examples{
tib <- tibble(node=c(1,2,3,4,5), name=c("root","inner node","leaf","leaf","leaf"))
count_leaves(tib)
}
