% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruning.R
\name{cross_validation}
\alias{cross_validation}
\title{Cross Validation}
\usage{
cross_validation(data, Lambda, m, type = NULL)
}
\arguments{
\item{data}{A list of training data.}

\item{Lambda}{Vector of possible lambda values}

\item{m}{Integer}

\item{type}{Either \code{"reg"} or \code{"class"}}
}
\value{
The optimal lambda estimation
}
\description{
Chooses optimal parameter lambda from vector \code{Lambda}
}
\details{
Choose the type of cross validation (regression or classification) which should be used.
See \code{cross_validation_reg} or \code{cross_validation_class} for further details.\cr
Return error if no valid type is chosen.
}
\examples{
data <- create_random_sample_data_reg(10,10)
Lambda <- c(1,2,3)/100
cross_validation(data, Lambda, m=3, type="reg")

}
