% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruning.R
\name{cross_validation_class}
\alias{cross_validation_class}
\title{Cross Validation (Classification)}
\usage{
cross_validation_class(data, Lambda, m)
}
\arguments{
\item{data}{A list of training data.}

\item{Lambda}{Vector of possible lambda values}

\item{m}{Integer}

\item{type}{Either \code{"reg"} or \code{"class"}}
}
\value{
The optimal lambda estimation
}
\description{
Chooses optimal parameter lambda from vector \code{Lambda}
}
\details{
The integer \code{m} lets you choose how many subsets the partition of the training data \code{data} should have.
For each complement of the subset and every component of \code{Lambda} pruning will be conducted. The function
returns the entry of \code{Lambda}, which minimizes the risk of the trees given through the pruninig process.
}
\examples{
data <- create_Sample_data_class(11)
Lambda <- c(1,2,3)/100
cross_validation_class(data, Lambda, m=3)

}
