% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruning.R
\name{cut_tree}
\alias{cut_tree}
\title{Cut tree}
\usage{
cut_tree(trees, new_leaf)
}
\arguments{
\item{trees}{A tibble containing the information of each node type of the CART in the colum "name".}

\item{new_leaf}{The node number where the tree will be cut and which will be turned from inner node into a leaf.}
}
\value{
A tibble in which all cut leaves are named "old leaf" and the former inner node \code{new leaf}, where the tree was cut, is named "leaf". \cr
All nodes, which where cut and hence turned into "old leaf", inherited the c_value from node number \code{new leaf}.
}
\description{
Cut a tree (CART) at a specific inner node value
}
\details{
\code{new leaf} has to be a positive whole number of an inner node of the tree. If \code{new leaf} is not a positive whole number
an error message will be returned. If \code{new leaf} is not an inner node, e.g. it is the node number of a leaf, or if it
is greater than he maximum node number in the tree, the function has no effect on the tree and will return the tree unchanged.
}
\examples{
tib <- tibble(node=c(1,2,3,4,5,6,7), name=c("root","inner node", "inner node", "leaf","leaf","leaf","leaf"), c_value=c(1,2,3,4,5,6,7))
tib
cut_tree(tib,3)

}
