% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GierigesVerfahren.R
\name{greedy_cart}
\alias{greedy_cart}
\title{Greedy Algorithm}
\usage{
greedy_cart(
  x,
  y,
  data,
  type = NULL,
  num_leaf = NULL,
  depth = NULL,
  num_split = 2,
  min_num = 1,
  m = 0
)
}
\arguments{
\item{x}{column/list name(s) of the x value(s)}

\item{y}{column/list name of the y value}

\item{data}{tibble or named list with data}

\item{type}{"reg" for regression tree\cr "class" for classification tree
\cr if \code{type} is missing the function tries to "guess" the type}

\item{num_leaf}{Condition to end: the tree has \code{num_leaf} leaves \cr must be greater than or equal to 1
\cr the default value is the maximal achievable number of leaves (the number of data points)}

\item{depth}{Condition to end: the tree has depth \code{depth}\cr must be greater than or equal to 0
\cr the default value is the maximal achievable depth}

\item{num_split}{split only nodes which contain at least \code{num_split} elements \cr must be greater than or equal to 2}

\item{min_num}{only split a node, if both child nodes have at least \code{min_num} elements \cr must be greater than or equal to 1}

\item{m}{parameter for Random Forest algorithm: positive number of coordinates which we want to use in each iteration
\cr the default value is the dimension of the data}
}
\value{
An environment with the elements \code{dim}, \code{values} and \code{tree}.\cr
\code{dim}: the dimension of the data. \cr
\code{values}: data in a tibble. \cr
\code{tree}: decision tree in the form of a tibble. It has the following columns \cr
\itemize{
\item \code{node}: node n has the child nodes 2n and 2n + 1
\item \code{name}: type of the node (root, inner node, leaf)
\item \code{split_index}: at which index we split the data
\item \code{split_point}: where we split the data. If node 2n has the split point s
then all data in this node is less than s. Analog if the node 2n + 1 has the split point s
then all data in this node is greater than or equal to s.
\item \code{y}: y values of data contained in this node
\item \code{A}: x values of data contained in this node
\item \code{c_value}: the approximate value for the data elements in a node
}
}
\description{
Greedy algorithm for either regression or classification data
}
\examples{
X <- runif(100,0,1)
e <- rnorm(100,0,0.2)
Y <- sin(2*pi*X) + e
data <- list(a = X, b = Y)
val <- greedy_cart(x = a, y = b, data = data, type = "reg")
val$values
val$tree

X1 <- runif(200,0,1)
X2 <- runif(200,0,1)
e <- rnorm(200,0,0.05)
k <- function(x,y) (x-0.5)*(y-0.5)
g <- function(x,y,e){
  Y <- c()
  for(i in seq_along(x)){
   if(k(X1[i],X2[i]) - e[i] <= 0){
     Y[i] <- 1
    } else{
      Y[i] <- 2
    }
  }
  Y
}
tbl <- tibble(x1 = X1, x2 = X2, y = g(X1,X2,e))
val <- greedy_cart(x = c(x1,x2), y = y, data = tbl, type = "class", depth = 3)
val$values
val$tree
}
