% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make_prediction}
\alias{make_prediction}
\title{Make Prediction
Make a Prediction of a datapoint or set of datapoints on a tree or set of trees.}
\usage{
make_prediction(tree_list, x_list, type = NULL)
}
\arguments{
\item{tree_list}{A list containing any number >=1 environments (trees) \cr as produced by \code{greedy_cart()}, \code{pruning()}, \code{bagging()}
or \code{random_forest()}}

\item{x_list}{A matrix of values to be predicted, where each column of \code{x_list} corresponds \cr to one datapoint.
\cr The number of rows corresponds to the dimension of the x values}

\item{type}{"reg" for regression trees\cr "class" for classification trees}
}
\description{
Make Prediction
Make a Prediction of a datapoint or set of datapoints on a tree or set of trees.
}
\examples{
treelist <- bagging(create_random_sample_data_reg_dim(5, 100, 3), 3, type = "reg")
xlist <- matrix(c(0.1, 0.4, 0.6, 0.2, 0.3, 0.6), ncol = 2)
predictions <- make_prediction(treelist$Bagged_Trees, xlist, "reg")
}
