% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruning.R
\name{partition}
\alias{partition}
\title{Partition}
\usage{
partition(data, m = 2)
}
\arguments{
\item{data}{A list of training data.}

\item{m}{Integer}
}
\value{
A list with the original data set and the complements of the subsets given by a partition with \code{m} subsets.
}
\description{
Takes a data set and creates the complements a partition of \code{m} subsets
}
\details{
The integer \code{m} lets you choose how many subsets the partition of the training data \code{data} should have.
Each complement of the subset will be returned as element of a list, with the first element being the original data set \code{data}.
Hence the list will have \code{m+1} entries. An error will be returned if \code{m} is greater than the number of
elements of \code{data}.
}
\examples{
X <- runif(50,0,1)
e <- rnorm(50,0,0.2)
Y <- sin(2*pi*X) + e
data <- list(x = matrix(X, nrow = 1), y = Y)
partition(data,m=3)

}
