% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotTree}
\alias{plotTree}
\title{Print Tree}
\usage{
plotTree(data)
}
\arguments{
\item{data}{an environment returned by the greedy_cart function (see ?greedy_cart) or a named list with a tree in form of a tibble and the name tree}
}
\value{
picture of a decision tree
}
\description{
Print a decision tree
}
\examples{
X <- runif(100,0,1)
e <- rnorm(100,0,0.2)
Y <- sin(2*pi*X) + e
data <- list(x = matrix(X, nrow = 1), y = Y)
val <- greedy_cart(x = x, y = y, data = data, type = "reg", num_leaf = 10)
plotTree(val)
plotTree(list(tree = val$tree))


}
