% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{printClassification}
\alias{printClassification}
\title{Plot (Classification)}
\usage{
printClassification(data, plotname)
}
\arguments{
\item{data}{an environment returned by the greedy_cart function (see ?greedy_cart)
\cr or a named list with an element named tree, which is the tree in tibble form and an element named values with the original
x and y values in the form a tibble with column names "x" and "y" for the x values and "classes" for the y values}

\item{plotname}{name which we want to give the plot}
}
\value{
a plot with the original values and the decision rule established by the tree
}
\description{
Plot the values and the decision rule established by a tree for classification data (in the two-dimensional case)
}
\examples{
X1 <- runif(200,0,1)
X2 <- runif(200,0,1)
e <- rnorm(200,0,0.05)
k <- function(x,y) (x-0.5)*(y-0.5)
g <- function(x,y,e){
  Y <- c()
  for(i in seq_along(x)){
   if(k(X1[i],X2[i]) - e[i] <= 0){
     Y[i] <- 1
    } else{
      Y[i] <- 2
    }
  }
  Y
}
tbl <- tibble(x = X1, y = X2, classes = g(X1,X2,e))
val <- greedy_cart(x = c(x,y), y = classes, data = tbl, type = "class", unique = TRUE, num_split = 20)
printClassification(val, "classification")
printClassification(list(tree = val$tree, values = tbl), "classification")
}
