% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{printRegression}
\alias{printRegression}
\title{Plot (Regression)}
\usage{
printRegression(data, plotname)
}
\arguments{
\item{data}{an environment returned by the greedy_cart function (see ?greedy_cart)
\cr or a named list with an element named tree, which is the tree in tibble form and an element with the name values containing the original
x and y values in the form a tibble with column names "x" and "y"}

\item{plotname}{name which we want to give the plot}
}
\value{
a plot with the original values and the decision rule established by the tree
}
\description{
Plot the values and the decision rule established by a tree for regression data (in the one-dimensional case)
}
\examples{
X <- runif(100,0,1)
e <- rnorm(100,0,0.2)
Y <- sin(2*pi*X) + e
data <- tibble(x = X, y = Y)
val <- greedy_cart(x = x, y = y, data = data, type = "reg", num_split = 10)
printRegression(val, "regression")
printRegression(list(tree = val$tree, values = data), "regression")


}
