% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruning.R
\name{pruning}
\alias{pruning}
\title{Cost-complexity Pruning}
\usage{
pruning(trees, lambda, type = NULL)
}
\arguments{
\item{trees}{A tibble containing the information of the tree.}

\item{lambda}{Parameter to weigh the amount of leaves per subtree}

\item{type}{Either \code{"reg"} or \code{"class"}}
}
\value{
The optimal subtree.
}
\description{
Chooses optimal subtree from CART for parameter \code{lambda}
}
\details{
Choose the type of pruning (regression or classification) which should be used.
See \code{pruning_regression} or \code{pruning_classification} for further details.\cr
Return error if no valid type is chosen.
}
\examples{
trees <- greedy_cart_classification(create_Sample_data_class())$tree
print(pruning(trees,1/80,type="class"),n=40)

trees <- greedy_cart_regression(create_random_sample_data_reg(10,20))$tree
print(pruning(trees,1/100,type="reg"),n=40)
}
