% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RForests.R
\name{random_forest}
\alias{random_forest}
\title{RANDOM FOREST ALGORITHM}
\usage{
random_forest(
  x,
  y,
  data,
  B,
  A,
  m,
  type = NULL,
  num_leaf = NULL,
  depth = 0,
  num_split = 2,
  min_num = 1
)
}
\arguments{
\item{data}{Tibble that contains Regression Data}

\item{m}{Number of Samples that should be aquired}

\item{t}{Condition to End: The Tree has t leaves}
}
\value{
tree in format: Tibble (BFS)
}
\description{
Condition to End: The Tree has t leaves - 6.52
}
