% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RForests.R
\name{random_forest_regression}
\alias{random_forest_regression}
\title{RANDOM FOREST ALGORITHM (Regression)}
\usage{
random_forest_regression(data, m, t)
}
\arguments{
\item{data}{Tibble that contains Regression Data}

\item{m}{Number of Samples that should be aquired}

\item{t}{Condition to End: The Tree has t leaves (insert tea leaf pun here)}
}
\value{
tree in format: Tibble (BFS)
}
\description{
Condition to End: The Tree has t leaves - 6.52
}
\examples{
random_forest_regression(create_random_sample_data_reg(10, 100), 30, 40)
}
