% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruning.R
\name{risk_tree_class}
\alias{risk_tree_class}
\title{Risk Tree Classification}
\usage{
risk_tree_class(trees)
}
\arguments{
\item{trees}{A tibble containing the information of the tree, e.g. y values of training data and c values for the estimations.}
}
\value{
The risk of the algorithm, which corresponds to the information in the tree.
}
\description{
Calculate the risk of a tree (CART) using the risk function for classification
}
\details{
The risk function for classification adds ones if the y value and its corresponding
c value estimation are not identical and divides that term by the number of training points.
}
\examples{
trees <- greedy_cart_classification(create_Sample_data_class())$tree
trees
risk_tree_class(trees)



}
