% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruning.R
\name{risk_tree_reg}
\alias{risk_tree_reg}
\title{Risk Tree Regression}
\usage{
risk_tree_reg(trees)
}
\arguments{
\item{trees}{A tibble containing the information of the tree, e.g. y values of training data and c values for the estimations.}
}
\value{
The risk of the algorithm, which corresponds to the information in the tree.
}
\description{
Calculate the risk of a tree (CART) using the risk function for regression
}
\details{
The risk function for regression sums the square of the difference of each y value of the training data and its corresponding
c value estimation and divides that term by the number of training points.
}
\examples{
trees <- greedy_cart_regression(create_realistic_sample_data_reg())$tree
trees
risk_tree_reg(trees)



}
