% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruning.R
\name{tree_seq_class}
\alias{tree_seq_class}
\title{Tree Sequencing (Classification)}
\usage{
tree_seq_class(trees)
}
\arguments{
\item{trees}{A tibble containing the information of the tree.}
}
\value{
A subtree which minimizes the risk of cutting the tree in relation to the number of leafs left.
}
\description{
Tree Sequencing used in cost-complexity pruning for classification CARTs
}
\details{
For each risk calculation of a tree the \code{risk_tree_class} function is used. The subtree, which will be returned,
minimizes the quotient of \code{(risk_tree_class(trees) - riks_tree_class(subtree)) / (count_leaves(subtree) - count_leaves(trees))} \cr
A column containing the risk of the subtree will be added to the tibble.
}
\examples{
trees <- greedy_cart_classification(create_Sample_data_class())$tree
print(tree_seq_class(trees),n=40)

}
