% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruning.R
\name{tree_seq_reg}
\alias{tree_seq_reg}
\title{Tree Sequencing (Regression)}
\usage{
tree_seq_reg(trees)
}
\arguments{
\item{trees}{A tibble containing the information of the tree.}
}
\value{
A subtree which minimizes the risk of cutting the tree in relation to the number of leafs left.
}
\description{
Tree Sequencing used in cost-complexity pruning for regression CARTs
}
\details{
For each risk calculation of a tree the \code{risk_tree_reg} function is used. The subtree, which will be returned,
minimizes the quotient of \code{(risk_tree_reg(trees) - riks_tree_reg(subtree)) / (count_leaves(subtree) - count_leaves(trees))} \cr
A column containing the risk of the subtree will be added to the tibble.

trees <- greedy_cart_regression(create_realistic_sample_data_reg())$tree
print(tree_seq_reg(trees),n=37)
}
