% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging.R
\name{bagging_classification}
\alias{bagging_classification}
\title{Bagging Algorithm (Classification)}
\usage{
bagging_classification(data, b = 5, pred_val = NULL)
}
\arguments{
\item{data}{a named list that contains classification data\cr the x values have the name x and are in
the form of a matrix \cr where the row number gives the dimension of the data\cr the y
values have the name y and are in the form of a vector}

\item{b}{Number of Bags to create. The Default Value is 5.}

\item{pred_val}{OPTIONAL: If wanting to predict some values in a matrix, each of length \cr equal one less than
the number of dimensions in our data set. \cr Else, input null.}
}
\value{
A named list with the elements \code{Bagged_Trees} and \code{Prediction}.\cr
\code{Bagged_Trees} is an unnamed list with a number of tree \cr elements (see ?greedy_cart_regression) equal to the input value \code{b}\cr
\code{Prediction} is the predicted class resulting from the \code{pred_val} data point.\cr If no \code{pred_val} was provided, this value is NULL
}
\description{
Bagging Algorithm for Classification data.
}
\examples{
bagging_classification(create_Sample_data_class(1), 4, matrix(c(0.1,0.4,0.6, 0.2, 0.4, 0, 1, 1), ncol = 4))
}
