% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging.R
\name{bagging_classification}
\alias{bagging_classification}
\title{RANDOM FOREST BAGGING ALGORITHM (Classification)}
\usage{
bagging_classification(data, b, x)
}
\arguments{
\item{data}{Tibble that contains Regression Data}

\item{x}{OPTIONAL: If wanting to predict a Value}

\item{m}{Number of Samples that should be aquired}
}
\value{
tree in format: Tibble (BFS)
}
\description{
Condition to End: The Tree has t leaves - 6.52
}
