% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging.R
\name{bagging_regression}
\alias{bagging_regression}
\title{RANDOM FOREST BAGGING ALGORITHM (Regression)}
\usage{
bagging_regression(data, b, x)
}
\arguments{
\item{data}{Tibble that contains Regression Data}

\item{b}{Number of Bags}

\item{x}{OPTIONAL: If wanting to predict a Value}
}
\value{
tree in format: Tibble (BFS)
}
\description{
Condition to End: The Tree has t leaves - 6.52
}
