% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GierigesVerfahren.R
\name{greedy_cart_classification}
\alias{greedy_cart_classification}
\title{GREEDY ALGORITHM (Classification)}
\usage{
greedy_cart_classification(data)
}
\arguments{
\item{data}{Tibble that contains Classification Data}
}
\value{
tree in format: Tibble (BFS)
}
\description{
Condition to End: Every Leaf contains one Value - 6.16
}
\examples{
greedy_cart_classification(create__random_sample_data_class(10, 50))
}
