% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GierigesVerfahren.R
\name{greedy_cart_regression}
\alias{greedy_cart_regression}
\title{GREEDY ALGORITHM (Regression)}
\usage{
greedy_cart_regression(data, depth = 0, num_split = 2, min_num = 1)
}
\arguments{
\item{data}{Tibble that contains Regression Data}
}
\value{
tree in format: Tibble (BFS)
}
\description{
Condition to End: Every Leaf contains one Value - 6.15
}
\examples{
greedy_cart_regression(create_random_sample_data_reg(12, 50))
}
