% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RForests.R
\name{random_forest_classification}
\alias{random_forest_classification}
\title{RANDOM FOREST ALGORITHM (Classification)}
\usage{
random_forest_classification(data, m, t)
}
\arguments{
\item{data}{Tibble that contains Regression Data}

\item{m}{Number of Samples that should be aquired}

\item{t}{Condition to End: The Tree has t leaves}
}
\value{
tree in format: Tibble (BFS)
}
\description{
Condition to End: The Tree has t leaves - 6.52
}
\examples{
random_forest_classification(create_random_sample_data_class(10, 100), 30, 40)
}
